refactor(core): extract runtime diagnostics sampler (#214)
Made-with: Cursor
